Merge all parts of evaluator into one single file
This is because there is a circular dependency issue that I can not
resolve now when doing exercise 4.16. The exercise requires
introducing a new procedure that depends on let and definition, these
packages in turn depends on procedure. However, this new procedure is
going to be used by procedure, thus a circular dependency.